+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add `Zo.cancel_all_orders{,_ix}`.
+
 ## [0.1.4] - 2022-08-02
 
 - Add `max_ts` parameter to `Zo.place_order`.