+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Add `Zo.cancel_all_orders{,_ix}`.
+- BREAKING: Add heimdall to `Zo.withdraw` accounts list.
 
 ## [0.1.4] - 2022-08-02